refactor project structure; downward jupiter package; migrate jupiter tool
