--- api/configuration.proto
@@ -1,10 +1,0 @@
-syntax = "proto3";
-package com.douyu.jupiter.proto.registry.configuration;
-
-option go_package="configuration";
-
-message Configuration {
-
-}
-
-message Route
