+++ cmd/jupiter/new/templates/.gitignore
