+++ cmd/jupiter/new/templates/cmd/main.go.tmpl
