+++ cmd/jupiter/new/templates/go.mod.tmpl
