+++ cmd/jupiter/new/templates/internal/app/model/db/user.go.tmpl
