+++ cmd/jupiter/new/templates/internal/app/model/init.go.tmpl
