+++ cmd/jupiter/new/templates/internal/app/service/init.go.tmpl
