+++ cmd/jupiter/new/templates/pb/hello/hello.proto
