+++ cmd/jupiter/new/templates/sql/user.sql
