+++ cmd/jupiter/new/tools.go
@@ -17,15 +17,16 @@ package new
 import (
 	"bytes"
 	"fmt"
-	"github.com/douyu/jupiter/pkg/util/xcolor"
-	"github.com/douyu/jupiter/tools/jupiter/common"
-	"github.com/gobuffalo/packr/v2"
-	"github.com/urfave/cli"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
 	"text/template"
+
+	"github.com/douyu/jupiter/cmd/jupiter/common"
+	"github.com/douyu/jupiter/pkg/util/xcolor"
+	"github.com/gobuffalo/packr/v2"
+	"github.com/urfave/cli"
 )
 
 // CreateProject create a template project for Jupiter
