--- pkg/application/init.go
@@ -1,1 +1,0 @@
-package application
