+++ pkg/application/option.go
@@ -1,0 +1,45 @@
+// Copyright 2021 rex lv
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package application
+
+import "github.com/douyu/jupiter/pkg/conf"
+
+type Option func(a *Application)
+
+type Disable int
+
+const (
+	DisableParserFlag      Disable = 1
+	DisableLoadConfig      Disable = 2
+	DisableDefaultGovernor Disable = 3
+)
+
+func (a *Application) WithOptions(options ...Option) {
+	for _, option := range options {
+		option(a)
+	}
+}
+
+func WithConfigParser(unmarshaller conf.Unmarshaller) Option {
+	return func(a *Application) {
+		a.configParser = unmarshaller
+	}
+}
+
+func WithDisable(d Disable) Option {
+	return func(a *Application) {
+		a.disableMap[d] = true
+	}
+}
