--- pkg/server/governor/http.go
@@ -1,58 +1,0 @@
-// Copyright 2020 Douyu
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package governor
-
-import (
-	"encoding/json"
-	"net/http"
-	"net/http/pprof"
-	"runtime/debug"
-)
-
-var (
-	// DefaultServeMux ...
-	DefaultServeMux = http.NewServeMux()
-	routes          = []string{}
-)
-
-func init() {
-	// 获取全部治理路由
-	HandleFunc("/routes", func(resp http.ResponseWriter, req *http.Request) {
-		json.NewEncoder(resp).Encode(routes)
-	})
-
-	HandleFunc("/debug/pprof/", pprof.Index)
-	HandleFunc("/debug/pprof/cmdline", pprof.Cmdline)
-	HandleFunc("/debug/pprof/profile", pprof.Profile)
-	HandleFunc("/debug/pprof/symbol", pprof.Symbol)
-	HandleFunc("/debug/pprof/trace", pprof.Trace)
-
-	if info, ok := debug.ReadBuildInfo(); ok {
-		HandleFunc("/modInfo", func(w http.ResponseWriter, r *http.Request) {
-			encoder := json.NewEncoder(w)
-			if r.URL.Query().Get("pretty") == "true" {
-				encoder.SetIndent("", "    ")
-			}
-			_ = encoder.Encode(info)
-		})
-	}
-}
-
-// HandleFunc ...
-func HandleFunc(pattern string, handler http.HandlerFunc) {
-	// todo: 增加安全管控
-	DefaultServeMux.HandleFunc(pattern, handler)
-	routes = append(routes, pattern)
-}
